Move hooks to own package
