+++ tests/compiler-hooks/build.gradle.kts
@@ -1,7 +1,7 @@
-import com.apollographql.apollo3.compiler.DefaultApolloCompilerKotlinHooks
 import com.apollographql.apollo3.compiler.codegen.ResolverKey
-import com.squareup.kotlinpoet.CodeBlock
+import com.apollographql.apollo3.compiler.hooks.DefaultApolloCompilerKotlinHooks
 import com.squareup.kotlinpoet.ClassName
+import com.squareup.kotlinpoet.CodeBlock
 import com.squareup.kotlinpoet.FileSpec
 import com.squareup.kotlinpoet.KModifier
 import com.squareup.kotlinpoet.TypeSpec
